Revert Speech-to-Text model back to 'long'
The 'short' model change was unverified — a clean-audio comparison
test showed no accuracy difference over 'long', and it was made
alongside the now-removed live-preview feature specifically to chase
a mismatch that stopped being visible once that feature was dropped.
Reverting to the last confirmed-working configuration.

+++ src/lib/googleStt.ts
@@ -49,10 +49,7 @@ export async function transcribeSpeech(audioBuffer: Buffer, languageCode: string
           config: {
             autoDecodingConfig: {},
             languageCodes: [languageCode],
-            // 'short' is tuned for brief, single-utterance audio (a few seconds) — matches this
-            // route's push-to-talk use case better than 'long', which targets continuous/longer
-            // recordings and can bias toward different word continuations on a short clip.
-            model: 'short',
+            model: 'long',
           },
           content: audioBuffer.toString('base64'),
         }),
